refactor: address CodeRabbit review feedback on PR #164
- Move pure algorithm functions to domain/utils/deduplication.ts
  (fixes architecture violation: skill layer importing from infrastructure)
- Fix --min-similarity 0 being silently dropped (parseFloat || null → Number.isNaN)
- Apply --since filter to conflict group query in dedupe
- Fix totalDuplicates to sum fact counts, not group counts
- Use Set for O(1) tagGroup membership checks instead of Array.includes()
- Add --group to CHANGELOG documentation
